Documentação Swagger + Esqueleto do projeto
