+++ jobby/src/main/java/com/jobby/g6/api/controller/CadastroController.java
@@ -1,0 +1,4 @@
+package com.jobby.g6.api.controller;
+
+public class CadastroController {
+}
