+++ jobby/src/main/java/com/jobby/g6/api/dto/CadastroDTO.java
@@ -1,0 +1,4 @@
+package com.jobby.g6.api.dto;
+
+public record CadastroDTO() {
+}
