+++ jobby/src/main/java/com/jobby/g6/domain/model/Endereco.java
@@ -1,5 +1,12 @@
-package com.jobby.g6.model;
+package com.jobby.g6.domain.model;
 
+import jakarta.persistence.Embeddable;
+import lombok.Getter;
+import lombok.Setter;
+
+@Embeddable
+@Getter
+@Setter
 public class Endereco {
 
     private Long cep;
