+++ jobby/src/main/java/com/jobby/g6/domain/repository/CadastroRepository.java
@@ -1,6 +1,6 @@
-package com.jobby.g6.repository;
+package com.jobby.g6.domain.repository;
 
-import com.jobby.g6.model.Cadastro;
+import com.jobby.g6.domain.model.Cadastro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CadastroRepository extends JpaRepository<Cadastro, Integer> {
