--- jobby/src/main/java/com/jobby/g6/model/Celular.java
@@ -1,7 +1,0 @@
-package com.jobby.g6.model;
-
-public class Celular {
-
-    private Long numero;
-    private boolean whatsapp;
-}
