--- jobby/src/main/java/com/jobby/g6/model/Cidade.java
@@ -1,9 +1,0 @@
-package com.jobby.g6.model;
-
-public class Cidade {
-
-    private Integer id;
-    private String nome;
-    private String estado;
-    private String sigla;
-}
