--- jobby/src/main/java/com/jobby/g6/model/Empresa.java
@@ -1,7 +1,0 @@
-package com.jobby.g6.model;
-
-public class Empresa {
-
-    private Integer id;
-    private String nome;
-}
