--- jobby/src/main/java/com/jobby/g6/model/PretensaoSalarial.java
@@ -1,7 +1,0 @@
-package com.jobby.g6.model;
-
-public class PretensaoSalarial {
-
-    private Double valorMinimo;
-    private Double valorMaximo;
-}
