--- jobby/src/main/java/com/jobby/g6/model/Profissao.java
@@ -1,7 +1,0 @@
-package com.jobby.g6.model;
-
-public class Profissao {
-
-    private Integer id;
-    private String nome;
-}
