+++ jobby/src/main/resources/application.properties
@@ -1,4 +1,7 @@
-spring.datasource.url= jdbc:postgresql://localhost:5432/jobby-g6
+spring.datasource.url=jdbc:postgresql://172.17.0.2:5432/jobby
 spring.datasource.username=postgres
 spring.datasource.password=lucarauj
 spring.jpa.hibernate.ddl-auto=update
+
+spring.jpa.show-sql=true
+#spring.flyway.locations=classpath:db/migration, classpath:db/testdata
